build: macOS/ARM platform compatibility fixes
- hardware.h: guard x86 SSE/AVX intrinsics, use arm64 equivalents
- pbgz_engine.cpp: pthread_setname_np signature for macOS
- reference.cpp: fseeko/ftello instead of fseeko64/ftello64, .c_str() in format strings
- path_util.cpp: minor macOS path fix
- fastq_actuator.cpp: static_cast to resolve Json::Value ambiguity
- CMakeLists.txt: link curl, lzma, CoreServices framework

+++ src/fastq_actuator.cpp
@@ -163,7 +163,7 @@ int32_t FastqCodecActuator::compress() {
     std::string md5;
     calcMd5sum(md5, inBlockPtr->getBuffer(), inBlockPtr->getDataLen());
     meta["md5"] = md5;
-    meta["idlines"] = inBlockPtr->getNpos().size() >> 2;
+    meta["idlines"] = static_cast<uint32_t>(inBlockPtr->getNpos().size() >> 2);
 
     // Compress meta information
     coder_json metaCoder;
